[cleanup] Renamed RepresentationChooser files
