+++ CMakeLists.txt
@@ -18,7 +18,7 @@ set(ADP_SOURCES
 	src/codechandler/ttml/TTML.cpp
 	src/common/AdaptiveStream.cpp
 	src/common/AdaptiveTree.cpp
-	src/common/RepresentationChooser.cpp
+	src/common/RepresentationChooserDefault.cpp
 	src/parser/DASHTree.cpp
 	src/parser/HLSTree.cpp
 	src/parser/SmoothTree.cpp
@@ -51,7 +51,7 @@ set(ADP_HEADERS
     src/common/AdaptiveDecrypter.h
 	src/common/AdaptiveStream.h
 	src/common/AdaptiveTree.h
-	src/common/RepresentationChooser.h
+	src/common/RepresentationChooserDefault.h
 	src/parser/DASHTree.h
 	src/parser/HLSTree.h
 	src/parser/SmoothTree.h
